+++ src/pages/auth/complete.tsx
@@ -1,0 +1,7 @@
+import type { NextPage } from 'next';
+
+const Complete: NextPage = () => {
+  return <div>complete</div>;
+};
+
+export default Complete;
