--- src/pages/auth/info.tsx
@@ -1,7 +1,0 @@
-import type { NextPage } from 'next';
-
-const Info: NextPage = () => {
-  return <div>info</div>;
-};
-
-export default Info;
